refac: align.c
