+++ includes/utils.h
@@ -3,6 +3,11 @@
 
 #include <stddef.h>
 
+/* align.c */
+size_t	align(size_t size, size_t align);
+size_t	align_size(size_t size);
+size_t align_chunk_size(size_t size);
+
 /* page_size.c */
 size_t	calc_map_size(size_t max_block_size);
