+++ srcs/malloc.c
@@ -7,24 +7,6 @@
 
 t_malloc	g_regions;
 
-size_t	align(size_t size, size_t align)
-{
-	return ((size + (align - 1)) & ~(align - 1));
-}
-
-size_t	align_size(size_t size)
-{
-	return align(size, BYTE);
-}
-
-size_t align_chunk_size(size_t size)
-{
-	if (size < MINSIZE)
-		size = MINSIZE;
-	size += HEADER_SIZE;
-	return align(size, MALLOC_ALIGNMENT);
-}
-
 void	*alloc_pages_by_size(size_t map_size, void *start)
 {
 	void	*ptr;
@@ -36,7 +18,6 @@ void	*alloc_pages_by_size(size_t map_size, void *start)
 	}
 	return (ptr);
 }
-
 
 void	init_region(t_region *region, e_size size_type)
 {
